### 2021-03-05 #49 @tawantayron
[Re] Context-Dependent Encoding of Fear and Extinction Memories in a Large-Scale Network Model of the Basal Amygdala
**Original article:** Vlachos, I., Herry, C., Lüthi, A., Aertsen, A., & Kumar, A. "[Context-dependent encoding of fear and extinction memories in a large-scale network model of the basal amygdala](https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1001104)". PLoS Comput Biol, v. 7, n. 3, p. e1001104, 2011.

**Scientific domain:** Computational Neuroscience
**Programming language:** Python and Brian 2
**Suggested editor:** @benoit-girard
**Suggested reviewer:** @mstimberg @heplesser @rgutzen @TiinaManninen
