### 2021-05-20 #49 @tawantayron
Dear editor,
We would like to kindly ask if there is any update on the progress of the review process of our submission.
We understand that we still are amidst difficult times and that the steps involving the revision of our manuscript can be longer than usual, but it would be great if this manuscript could be progressed in the not too distant future.
Cordially,
