### 2021-05-26 #49 @eroesch
Hi @tawantayron, I am running into a runtime error, running runsim.py with protocol=1, =2 (I am still running 3 and above, some warnings but so far so good). See below. The code seems to carry on in a loop, and I assumed the code might still be running. After 24h+ of running though, it still does not complete. Can you let me know what you think?

ERROR      Brian 2 encountered an unexpected error. If you think this is a bug in Brian 2, please report this issue either to the discourse forum at <http://brian.discourse.group/>, or to the issue tracker at <https://github.com/brian-team/brian2/issues>. Please include this file with debug information in your report: /var/folders/8f/5f00v5l94r91mtfll9t0x2sm0000gn/T/brian_debug_anijad72.log  Additionally, you can also include a copy of the script that was run, available at: /var/folders/8f/5f00v5l94r91mtfll9t0x2sm0000gn/T/brian_script_54_70__1.py Thanks! [brian2]
Traceback ....

RuntimeError:
        An attempt has been made to start a new process before the
        current process has finished its bootstrapping phase.

        This probably means that you are not using fork to start your
        child processes and you have forgotten to use the proper idiom
        in the main module:

            if __name__ == '__main__':
                freeze_support()
                ...

        The "freeze_support()" line can be omitted if the program
        is not going to be frozen to produce an executable.
