### 2021-05-26 #49 @eroesch
Actually I spoke to soon, protocol=3 works, but not 4 or 5.
