### 2021-05-31 #49 @tawantayron
Dear editor,

We have double-checked the code on Linux systems and fix small issues encountered. We also noted a bug when running on MacOS that should be solved now, in case that is the system you are using let us know if you run into more issues while executing the code.
