### 2021-05-31 #49 @niltonlk
Dear @eroesch,
The issue was that in MacOS, the code block that creates the pool (`p = multiprocessing.Pool(processes=n_procs)`) and calls `p.map(func, arg_list)` to run some simulations in parallel (with different parameters), should be enclosed in an if clause (`if __name__ == '__main__':`) to avoit that the code is called by every process (codes that reads the simulation data and plots the results were also included withing the same if clause). This was not necessary on Linux, which is the environment we have originally tested.
