### 2021-09-20 #49 @rougier
@eroesch Any update ? Also, maybe @benoit-girard would be interested in reviewing this submission on a Large-Scale Network Model of the Basal Amygdala
