### 2021-09-20 #49 @eroesch
Hi! I was indeed off the grid for a while! I had issues running the original code, and haven't had time to try all of @niltonlk's fixes. I have timeblocked time to run through it this week. @benoit-girard, Would you mind giving it a go as well? What would be a good timeline for you?
