### 2021-10-04 #49 @eroesch
Jep. All done. It took a few days to run the last models, sorry. @benoit-girard, I do not think I need you to do a review.

Thank you for fixing the code, I can now confirm that I have managed to run all your scripts, and have reproduced all the figures presented in your submission; the more complex models took a few days. The figures you produce are similar to those in the original work (but not exactly the same), and the equations you have had to implement are close enough to the original ones. I am happy to recommend acceptance. I do have some minor recommendations. Please let me know if you are able to do this; if not I can make some edits myself. @rougier, I am going ahead with the publishing workflow. @tawantayron, do let me know how long you think you'd need.

TODO:
-  Please provide a requirement.txt file containing the required libraries and packages, and their versions that need to be used to run the models. Your README.md is quite verbose, and it would ease the workflow for someone else to reproduce your environment.
- You do a great job describing the missing parameters and the guesswork you have had to do, but you do not follow strictly the way to present the equations from the original paper, which can be confusing for someone who is not from the field. At first glance and for the untrained eye, they do look quite different: particularly, your equations 1-3 and the original 1-2, and your 5 to the original 5. At a minimum, add sentences pointing out the differences and explaining why they are equivalent (e.g. mentioning the equivalence of notation, and explaining the inclusion of the conductance in Eq 5). I do not necessarily think you should rename your variables, but it would help.
- Make sure the captions for all your figures mention the original figure it refers to. Obviously, in the core text, explain any differences. e.g. Your figure 7 mentions the original Figure 5, but really refers to the original Figure 7, showing the renewal scenario. That is not clear.
