### 2021-10-11 #49 @tawantayron
Dear editor @eroesch , we made the following modifications:

- We added the file “requirements.txt” which contains all versions of the packages used to run the models, and also added simplified commands to install an anaconda environment with all the packages needed to run the codes. We also create a document with the instructions to install the packages using pip.
- Thanks to the editor for highlighting this point. We have added brief explanations to the manuscript in order to clarify that the dynamics equations of the models as we wrote are equivalent to the equations used in the original article.
- We reviewed all the figures in the manuscript, clarifying which ones correspond to the reference article and which ones are additional figures that add details to the studied models.

We are at your disposal for any other queries. Cordially,
