### 2021-11-09 #49 @eroesch
All done. I think: [https://zenodo.org/record/5657320#.YYp2_r2mOAk](https://zenodo.org/record/5657320#.YYp2_r2mOAk).
I am not quite sure I understand how to update the website though, @rougier...

Also, I have had to update the metadata myself and latex the pdf (first because I needed to add the publication date, and then because I have had to re-reserve a doi), then I also hit a bump with the sandbox in step 2 (probably because I wasn't using the appropriate version of PyYAML). All sorted in the end, with a bit of manual sweat.
I am closing the issue.
