### 2021-11-22 #49 @rougier
@eroeach Sorry for late answer. To update the website, you just need to add the bibtex entry (using yaml to bibtex) to the [published.bib](https://github.com/ReScience/rescience.github.io/blob/sources/_bibliography/published.bib) file.
